move compare function to framework
